[WIP] Updating the parser to do dynamic kw def
Up til now, a function is called using the "call" keyword. I'm ok
with verbosity because the programs are designed for language models,
but this is taking it a little bit too far in the face of a relatively
elegant (I think) solution.

Right now, a call looks like,

call my_function
  arg a
  arg b
  arg c
end

What I want to do is allow for

my_function
  arg a
  arg b
end

I think my primary motivation is that it actually makes the language
even more "regular" or homogenous. It means that we can use user-
defined keywards exactly the same way as we can use default keywords
(I think). Of course, I don't really know what I'm talking about,
but the implementation for this feels like it's pushing the project
in the right direction. Basically, we're adding more fields to the
Expression struct (an expression represents a node in the parse
tree).

It's adding a little bit of weirdness to the parser code because
the parser has to keep track of the names that are in scope at parse
time whereas before the parser was more of a "static" operation.
I'm ok with the weirdness because instrumenting the parser is kind
of one of the primary points of the project. I donno.