Big "system 2" kind of refurb
After hacking on the language for a while, I felt like I had gotten
to a point where I knew how simple I wanted it and I also knew what
code I needed to to make it happen, so I decided to dive into
something like a rewrite. So far, the parser is done. The gist
of the design is basically that the language has

- function definitions
- function calls
- literals
- calculations

and nothing else. There's probably a better or standard name for
"calculation" here, I'm just using it to mean "stuff with operators
and without blocks". I didn't use the word expression because I'm
already using expression to mean "any thing that gets evald"