[Broken] Saving for posterity
Sooooooo I think these changes are not great. I basically wanted
everything to be a lambda so that there's exactly one way of calling
functionality, but it ended up introducing more complexity rather
than simplifying things. I'm going to try to articulate why I've
gotten stuck so that I can learn from my mistake.

So a function's implementation can either (1) take it's closure and
then evaluate its children in some particular way and then use the
results in some particular way or (2) have it's arguments passed
directly to it after they've been evaluated. (2) doesn't work for
conditional execution and (1) makes it hard to pass values that
don't come directly from evaluating the parse tree. An example of
"values that don't come from the parse tree" would be the arguments
in, say, an Array.map callback.

So in order to implement "For" in the (2) regime, you have to inject
a bunch of expressions into the parse tree. Actually maybe this is
fine? I tried to do this as eval time but maybe it makes sense to do
it in the parse tree?...

Ahh, probably not because you stil have to artificially declare the
names of the parameters to the callback. Let me think for a second
if we want to do this even though it's a pain...

So one way we could do it is by adding a side channel for name
resolution that doesn't use the tree's names...

Yeah I don't know. This is starting to feel strongly like a case
of "just have 2 kinds of things". GG.