Add conditional trajectory expansion
Before, every eval function expanded an expressions trajectory by each
of the expression's children. Now, conditional evaluators only expand
when that expression is actually going to be evaluated.

One side effect of this is that the trajectory tree's structure is
different from the expression's, it can have less branching. This is
fine programmatically because every trajectory node points to its
corresponding expression, but it might make visual debugging harder.