WIP Named parameters
Right now I'm trying to shim in named parameters but I'm not taking
the time to do it right. It's turning into a situation where the
slow way is the fast way, so I'm going to do that tomorrow. Let's
think about the situation:

- some functions have mapped arguments
  - some of these have control flow considerations
  - some of these don't
- some functions have array arguments
  - some of these have control flow considerations
  - some of these don't

...

You know what, I've thought about it, and I think I've lost sight
of my original goal. I'm trying to make the syntax and grammar to
pleasant for __myself__ when I know that the rasion d'etre is to
make the syntax and grammar extremely regular and programmable.